Characters page (#14)
* Switch to i64

* More usize to i64 fixes

* Fix missing character exposure

* Refactoring for cleaner GraphQL

* Add ugly, working characters page

* Character images

* Cargo fmt fixes

* First draft of character sheet look and feel

+++ ui/src/character_sheet/attacks.rs
@@ -3,7 +3,7 @@ use yew::{html, Component, ComponentLink, Html, Properties, ShouldRender};
 #[derive(Clone)]
 pub struct Attack {
     pub name: String,
-    pub bonus: usize,
+    pub bonus: i64,
     pub damage: String,
 }
